requirements.txt
transformers
peft
bitsandbytes
accelerate
tensorflow
torch
numpy
timm
einops
pillow
pynput
psutil
python-pptx
python-docx
selenium
fastapi
uvicorn
opencv-python
pyautogui
pytesseract
openai-whisper
ollama
pydantic
python-multipart
requests
